+++ veewee/scientificlinux-6.5-x86_64-netboot/chef.sh
@@ -1,2 +1,0 @@
-# Install Chef
-gem install --no-ri --no-rdoc chef