+++ veewee/scientificlinux-6.5-x86_64-netboot/ruby.sh
@@ -1,3 +1,4 @@
-# Install Ruby
+
+# Install Ruby, Gem and Ruby devel packages.
 yum -y install ruby ruby-devel rubygems