+++ veewee/scientificlinux-6.5-x86_64-netboot/vagrant.sh
@@ -1,10 +1,12 @@
+# Setup VM for Vagrant.
+
 # Vagrant specific
 date > /etc/vagrant_box_build_time
 
 # Add vagrant user
 /usr/sbin/groupadd vagrant
 /usr/sbin/useradd vagrant -g vagrant -G wheel
-echo "vagrant"|passwd --stdin vagrant
+echo "vagrant" | passwd --stdin vagrant
 echo "vagrant        ALL=(ALL)       NOPASSWD: ALL" >> /etc/sudoers.d/vagrant
 chmod 0440 /etc/sudoers.d/vagrant