+++ veewee/scientificlinux-6.5-x86_64-netboot/zerodisk.sh
@@ -1,3 +1,8 @@
+
 # Zero out the free space to save space in the final image:
 dd if=/dev/zero of=/EMPTY bs=1M
 rm -f /EMPTY
+
+# Remove veewee postinstall files.
+rm -f /home/veewee/*.sh
+