+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 happy-deployer CHANGELOG
 =====================
 
+0.2.4
+-----
+* Updated SL6 veewee template.
+* Updated and fixed manifests.
+* Add new puppet recipe for installing jekyll.
+
 0.2.2 - 0.2.3
 -----
 * Fixed PostgreSQL user password file.